refactor: flatten Checkout views structure and consolidate types
