+++ src/views/Checkout/views/index.ts
@@ -1,5 +1,5 @@
-export { default as CheckoutCancelView } from './CheckoutCancelView/CheckoutCancelView';
-export { CheckoutSessionId } from './CheckoutSession/CheckoutSessionId';
-export { default as CheckoutSuccessView } from './CheckoutSuccessView/CheckoutSuccessView';
-export { default as PcCloudSuccess } from './CheckoutSuccessView/PcCloudSuccess';
-export { default as CheckoutViewWrapper } from './IntegratedCheckoutView/CheckoutViewWrapper';
+export { default as CheckoutCancelView } from './CheckoutCancelView';
+export { CheckoutSessionId } from './CheckoutSessionId';
+export { default as CheckoutSuccessView } from './CheckoutSuccessView';
+export { default as PcCloudSuccess } from './PcCloudSuccess';
+export { default as CheckoutViewWrapper } from './CheckoutViewWrapper';
